requirements.txt
google-api-python-client==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
pandas==1.3.3
gspread==4.0.1
tqdm==4.54.1